services are now only bootstrapped on desktop platforms
